+++ CHANGELOG.md
@@ -28,10 +28,22 @@ Read our [guidelines for writing a good changelog entry](https://github.com/biom
 
 - Add support for `BIOME_BINARY` environment variable to override the location of the binary. Contributed by @ematipico
 
+- Add option `--indent-width`, and deprecated the option `--indent-size`. Contributed by @ematipico
+- Add option `--javascript-formatter-indent-width`, and deprecated the option `--javascript-formatter-indent-size`. Contributed by @ematipico
+- Add option `--json-formatter-indent-width`, and deprecated the option `--json-formatter-indent-size`. Contributed by @ematipico
+
 #### Enhancements
 - Deprecated the environment variable `ROME_BINARY`. Contributed by @ematipico
 
 ### Configuration
+
+#### New features
+
+- Add option `formatter.indentWidth`, and deprecated the option `formatter.indentSize`. Contributed by @ematipico
+- Add option `javascript.formatter.indentWidth`, and deprecated the option `javascript.formatter.indentSize`. Contributed by @ematipico
+- Add option `json.formatter.indentWidth`, and deprecated the option `json.formatter.indentSize`. Contributed by @ematipico
+
+
 ### Editors
 ### Formatter